Django DB added
Because MAL's API sucks, we have to maintain a backend database in order to have easy access to a master list of anime and their associated data. Currently, MAL's API doesn't allow search by genre, only keyword.

I added an AnimeEntry class model that (currently) saves a show's name, rank, MAL ID, and a base64 encoded string of the list of genres. I did this for simplicity, although there may be better ways to do it.

When the updateDB function is called, a given number of instances of this class are added to the backend database. Old data is currently being overwritten each time it's called.

There hasn't been any frontend functionality added, but I think this is a step in that direction. What needs to be done now is checking the database against the user's MAL list to find recommendations, as well as a refactor (way too many free functions). I leave those as exercises for the reader.