DB Restructure, basic recommendation
Rewrote jank genre database storage to use a ManyToMany relationship instead of JSON/base64 encoding. It's pretty expensive to run updates, but queries based on genre are actually possible now. With the old system, we'd have to exhaustively decode and store in memory all the JSON data to be able to find specific shows based on genre.

One minor change is raising the user result search limit to 1000 (in order to get a full picture of what they've watched to avoid duplicates), as well as a basic demo of genre-matching recommendations.

No frontend work has been done, and code still needs a refactor into object oriented-whatever.