non mal user support
users can now give:
- a username, no genres
- a username, genres
- no username, genres
- no username, no genres

in the first case, the algorithm determines top anime genres.
In the second, the algorithm uses both what it's given and what it auto determines. user provided search options are prioritized
In the third, the algorithm skips auto genres and just uses what it's given.
in the last case, I just have the algorithm search up a default list of genres