maximum results option
yup. Enter a number and get that number of results. defaults to 20 if nothing is given.

also i cleaned up the code and the tweaked the interface an imperceptible bit (the word "optional" appears as a hideable placeholder)

also, idk if I broke it or if it has been, but the home bar disappears upon executing a search. oops?